fix: final_score를 REAL로 명시적 캐스팅
- final_score가 DOUBLE PRECISION으로 계산되므로 ::REAL 캐스팅
- 42804 타입 불일치 오류 해결 (컬럼 14)

+++ supabase/migrations/create_personalized_feed_function.sql
@@ -114,7 +114,7 @@ BEGIN
     fr.views_count,
     fr.hot_score,
     fr.view_count::INTEGER,
-    fr.final_score,
+    fr.final_score::REAL,
     fr.adjusted_priority AS sort_priority
   FROM final_ranked fr
   ORDER BY
